==> 01_diabetes/02_diabetes_scripts/02_diabetes_clean.Rmd <==
---
title: "Diabetes - Data Cleaning"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---

# Clean Diabetes Dataset

## Goal

After executing this script, the diabetes dataset should be cleaned and ready for modelling in Python.

## Install packages

```{r}
pacman::p_load(tidyverse, 
               here,
               naniar,
               reticulate,
               wesanderson)
```

## Load data

```{r}
diab_data_raw = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                              "01_diabetes_data_load.rds"))
```

## Glimpse on the data

### Overview over data

```{r}
diab_data_raw
```

```{r}
str(diab_data_raw)
```

### Check for Missing Values

Table of absolute missing values per variable as well as percentage of missing values per variable

```{r}
miss_var_summary(diab_data_raw)
```

Plot with most frequent missingn patterns with according absolute frequencies

```{r}
gg_miss_upset(diab_data_raw) 
```

## Cleaning

- delete patient_nbr and make encounter_id the rowname 
- delete weight, payer_code and medical_specialty, examide, citoglipton, metformin-rosiglitazone, metformin-pioglitazone
- delete incomplete cases (wrt race, diag_1, diag_2, diag_2) = 3713 cases
- delete cases with "no" in readmitted variable (no record of readmission) = 52338 cases
=> 45,715 cases in final dataset

- change readmitted into classes (with "<30" as positive class) -> rate of 1:3.13 (p:n)?
- change data type appropriately (factors with python?)

```{r}
diab_data_clean = diab_data_raw %>% 
  arrange(encounter_id) %>% 
  column_to_rownames(var = "encounter_id") %>% 
  select(!c(patient_nbr, 
            weight, 
            payer_code, 
            medical_specialty,
            examide,
            citoglipton,
            diag_1,
            diag_2,
            diag_3,
            "metformin-rosiglitazone",
            "metformin-pioglitazone")) %>% 
  drop_na() %>% 
  filter(readmitted != "NO") %>% 
  mutate(across(.cols = c(race,
                          gender,
                          admission_type_id,
                          discharge_disposition_id,
                          admission_source_id,
                          # diag_1,
                          # diag_2,
                          # diag_3,
                          max_glu_serum,
                          A1Cresult,
                          metformin,
                          repaglinide,
                          nateglinide,
                          chlorpropamide,
                          glimepiride,
                          acetohexamide,
                          glipizide,
                          glyburide,
                          tolbutamide,
                          pioglitazone,
                          rosiglitazone,
                          acarbose,
                          miglitol,
                          troglitazone,
                          tolazamide,
                          insulin,
                          "glyburide-metformin",
                          "glipizide-metformin",
                          "glimepiride-pioglitazone",
                          change,
                          diabetesMed
                          ),
                .fns = ~ factor(.)
                )
         ) %>% 
  mutate(age = factor(age, ordered = TRUE)) %>% 
  mutate(readmitted = ifelse(readmitted=="<30", 1, 0))
```

## Exploratory Data Analysis
### Dataset overview
```{r}
diab_data_clean
```

```{r}
str(diab_data_clean)
```

### Visualization of the class variable (readmitted) & the protected variables (race, age & gender)

#### Definition of colors
```{r}
colors_discrete = rep(wes_palette("Darjeeling1"), 3)
```

#### Class variable: Readmitted
```{r}
diab_eda_class = diab_data_clean %>% 
  select(readmitted) %>% 
  mutate(readmitted = ifelse(readmitted == 1, "<30", ">30")) %>% 
  ggplot(aes(x = fct_rev(fct_infreq(readmitted)), fill = fct_rev(fct_infreq(readmitted)))) +
  geom_bar() +
  labs(title = "Class Frequencies in the Diabetes Dataset",
       subtitle = "With Class Variable 'Readmitted'",
       x = "Days until Readmission", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none")

diab_eda_class
```

```{r}
ggsave("01_diab_eda_class.png",
       plot = diab_eda_class,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 25,
       height = 15,
       units = c("cm")
       )
```

#### Protected variable: Race
```{r}
diab_eda_race = diab_data_clean %>% 
  select(race) %>% 
  ggplot(aes(x = fct_infreq(race), fill = fct_infreq(race))) +
  geom_bar() +
  labs(title = "Race Frequencies in the Diabetes Dataset",
       x = "Race", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none")

diab_eda_race
```

```{r}
ggsave("02_diab_eda_race.png",
       plot = diab_eda_race,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 25,
       height = 15,
       units = c("cm")
       )
```

#### Protected variable: Age
```{r}
diab_eda_age = diab_data_clean %>% 
  select(age) %>% 
  ggplot(aes(x = age, fill = age)) +
  geom_bar() +
  labs(title = "Frequencies of Age Groups in the Diabetes Dataset",
       x = "Age Group", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none") 

diab_eda_age
```

```{r}
ggsave("03_diab_eda_age.png",
       plot = diab_eda_age,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 25,
       height = 15,
       units = c("cm")
       )
```

#### Protected variable: Gender
```{r}
diab_eda_gender = diab_data_clean %>% 
  select(gender) %>% 
  ggplot(aes(x = fct_infreq(gender), fill = fct_infreq(gender))) +
  geom_bar() +
  labs(title = "Gender Frequencies in the Diabetes Dataset",
       x = "Gender", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none")

diab_eda_gender
```

```{r}
ggsave("04_diab_eda_gender.png",
       plot = diab_eda_gender,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 15,
       height = 15,
       units = c("cm")
       )
```

#### Protected variables: Race & Age
```{r}
diab_eda_race_age = diab_data_clean %>% 
  select(race, age) %>% 
  ggplot(aes(x = age, fill = age)) +
  geom_bar() +
  facet_wrap(~race) +
  labs(title = "Frequencies of Age Groups in the Diabetes Dataset",
       subtitle = "Grouped by Race",
       x = "Age Group", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none") +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

diab_eda_race_age
```

```{r}
ggsave("05_diab_eda_race_age.png",
       plot = diab_eda_race_age,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 25,
       height = 15,
       units = c("cm")
       )
```

#### Protected variables: Race & Gender
```{r}
diab_eda_race_gender = diab_data_clean %>% 
  select(race, gender) %>% 
  ggplot(aes(x = race, fill = race)) +
  geom_bar() +
  facet_wrap(~gender) +
  labs(title = "Race Frequencies in the Diabetes Dataset",
       subtitle = "Grouped by Gender",
       x = "Race", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none") +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

diab_eda_race_gender
```

```{r}
ggsave("06_diab_eda_race_gender.png",
       plot = diab_eda_race_gender,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 25,
       height = 15,
       units = c("cm")
       )
```

#### Protected variables: Age & Gender
```{r}
diab_eda_age_gender = diab_data_clean %>% 
  select(age, gender) %>% 
  ggplot(aes(x = age, fill = age)) +
  geom_bar() +
  facet_wrap(~gender) +
  labs(title = "Race Frequencies in the Diabetes Dataset",
       subtitle = "Grouped by Gender",
       x = "Race", 
       y = "Count") +
  scale_fill_manual(values = colors_discrete) +
  guides(fill = "none") +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

diab_eda_age_gender
```

```{r}
ggsave("07_diab_eda_age_gender.png",
       plot = diab_eda_age_gender,
       device = "png",
       path = here("01_diabetes", "03_diabetes_outputs", "03_diabetes_outputs_graphs"),
       dpi = 300,
       width = 25,
       height = 15,
       units = c("cm")
       )
```


## Save data

```{r}
saveRDS(diab_data_clean, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "02_diabetes_data_clean.rds")
        )
```
