==> 01_diabetes/02_diabetes_scripts/03_diabetes_model_preperation.Rmd <==
---
title: "Diabetes - Model Preparation"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---
# Preparation of the Diabetes Dataset for Modeling

## Goal

After executing this script, the diabetes dataset should be ready to be modeled as deep neural network in keras.

## Preperations

### Load R packages

```{r}
pacman::p_load(tidyverse, 
               here,
               reticulate)
```

### Activate Python environment

```{r}
reticulate::use_condaenv(condaenv = "Masterarbeit", 
                         required = TRUE
                         )
```

### Import Python packages

```{python}
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.compose import make_column_transformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import make_column_selector
from sklearn.model_selection import train_test_split
from sklearn import preprocessing
```


### Load data

```{r}
diab_data_clean = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                                      "02_diabetes_data_clean.rds"))
```

### Transform to python data frame

```{python}
diab_data_clean_pre = r.diab_data_clean
```

### Function for Standardization of the numeric features & One-Hot-Encoding of the categorical features

```{python}
# function for cleaning column names of add-ons from transformation steps (help function for feature_transformator below)
def clean_colnames(colnames):
  
  # delete/replace some specific strings
  clean_names = list(map(lambda x: x.replace("remainder__", "").replace("onehotencoder__", "").replace("standardscaler__", "").replace("-", "_").replace(" ", "_").replace("[", "").replace(")", ""), colnames))
  
  return(clean_names)



# actual transformation function
def diab_feature_transformator(df):
  
  # define transformation for each data type
  column_trans = make_column_transformer(
    
      (OneHotEncoder(), make_column_selector(dtype_exclude = float)),      # dummy-encoding of categorical features
      
      (StandardScaler(), make_column_selector(pattern = "^(?!readmitted$).*$",dtype_include = float)),     # standardization of numerical features except target feature
      
      remainder = 'passthrough'   # target feature remains in its original binary state
    )
  
  # create new dataframe with transformed features and column names 
  new_df = pd.DataFrame(column_trans.fit_transform(df).toarray(),
                        index = df.index,
                        columns = clean_colnames(column_trans.get_feature_names_out())
                        )
  return(new_df)

```

### Feature Transformation 

```{python}
diab_data_clean = diab_feature_transformator(diab_data_clean_pre)
```


## Preparation of the dataset

### Ratios for training, testing and validation set
```{python}
diab_train_ratio = 0.75
diab_validation_ratio = 0.15
diab_test_ratio = 0.1
```


### Train-test-split for training and testing datasets
```{python}
diab_X_train_raw, diab_X_test_raw, diab_y_train_raw, diab_y_test_raw = train_test_split(diab_data_clean.iloc[:,:-1], 
                                                                                        diab_data_clean.iloc[:,-1], 
                                                                                        test_size=diab_test_ratio,
                                                                                        random_state=42,
                                                                                        shuffle = True)
```

### Train-test-split for training and validation datasets
```{python}
diab_X_train_raw, diab_X_val_raw, diab_y_train_raw, diab_y_val_raw = train_test_split(diab_X_train_raw, 
                                                                                      diab_y_train_raw, 
                                                                                      test_size=diab_validation_ratio/(1-diab_test_ratio),
                                                                                      random_state=24,
                                                                                      shuffle = True)
```

### Reset all indices
```{python}
diab_X_train = diab_X_train_raw.reset_index(drop=True)
diab_y_train = diab_y_train_raw.reset_index(drop=True)

diab_X_val = diab_X_val_raw.reset_index(drop=True)
diab_y_val = diab_y_val_raw.reset_index(drop=True)

diab_X_test = diab_X_test_raw.reset_index(drop=True)
diab_y_test = diab_y_test_raw.reset_index(drop=True)
```

### Save column names for interpretation
```{python}
diab_col_names_enc = diab_X_train.columns.values.tolist()
```

```{python}
diab_X_train
diab_y_train

diab_X_val
diab_y_val

diab_X_test
diab_y_test

diab_col_names_enc
```

### Save as RDS
```{r}
diab_X_train = py$diab_X_train
saveRDS(diab_X_train, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "03_diabetes_data_model_X_train.rds")
        )

diab_y_train = py$diab_y_train
saveRDS(diab_y_train, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "04_diabetes_data_model_y_train.rds")
        )

diab_X_val = py$diab_X_val
saveRDS(diab_X_val, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "05_diabetes_data_model_X_val.rds")
        )

diab_y_val = py$diab_y_val
saveRDS(diab_y_val, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "06_diabetes_data_model_y_val.rds")
        )

diab_X_test = py$diab_X_test
saveRDS(diab_X_test, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "07_diabetes_data_model_X_test.rds")
        )

diab_y_test = py$diab_y_test
saveRDS(diab_y_test, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "08_diabetes_data_model_y_test.rds")
        )

diab_col_names_enc = py$diab_col_names_enc
saveRDS(diab_col_names_enc, 
        file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                    "09_diabetes_data_model_col_names_enc.rds")
        )
```
