==> 01_diabetes/02_diabetes_scripts/04_diabetes_model.Rmd <==
---
title: "Diabetes - Modelling"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---
# Model Diabetes Dataset

## Goal

After executing this script, a deep neural network for the diabetes dataset should be modeled.

## Preperations

### Load R packages

```{r}
pacman::p_load(tidyverse, 
               here,
               reticulate)
```

### Activate Python environment

```{r}
reticulate::use_condaenv(condaenv = "Masterarbeit", 
                         required = TRUE
                         )
```

### Import Python packages

```{python}
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import tensorflow as tf
```

### Load data
```{r}
diab_X_train_pd = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                              "03_diabetes_data_model_X_train.rds"))

diab_y_train_pd = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                              "04_diabetes_data_model_y_train.rds"))

diab_X_val_pd = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                            "05_diabetes_data_model_X_val.rds"))

diab_y_val_pd = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                            "06_diabetes_data_model_y_val.rds"))

diab_X_test_pd = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                             "07_diabetes_data_model_X_test.rds"))

diab_y_test_pd = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                             "08_diabetes_data_model_y_test.rds"))

diab_col_names_enc = readRDS(file = here("01_diabetes", "01_diabetes_data", "02_diabetes_data_processed",
                                    "09_diabetes_data_model_col_names_enc.rds"))
```

### Transfer dataframes to python
```{python}
diab_X_train_pd = r.diab_X_train_pd
diab_y_train_pd = r.diab_y_train_pd

diab_X_val_pd = r.diab_X_val_pd
diab_y_val_pd = r.diab_y_val_pd

diab_X_test_pd = r.diab_X_test_pd
diab_y_test_pd = r.diab_y_test_pd

diab_col_names_enc = r.diab_col_names_enc
```
